Added org mode with deft
Org mode is suitable as a terminal based note taking tool. I am picking this
over vim because of the summarization and outlining support this adds to my
workflow. It also becomes possible to have a single respositry across the
workweek.

Exploring deft; this is supposed to be like notational velocity.